refactor: extract bundle subelement hierarchy
